Add LICENSE file for qtpropertybrowser
Add separate LICENSE file for the bundled qtpropertybrowser library to
make binary packaging in distributions like Fedora Linux easier. The
license is already included at the beginning of all source files, but
this is not enough because the used BSD 3 Clause license requires
including a copy of the license with the binary. On the other hand, a
separate file with only qtpropertybrowser's license text is easy to
package with the binary.